+++ container/openejb-corba-integration/pom.xml
@@ -26,14 +26,20 @@
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <packaging>pom</packaging>
+  <artifactId>openejb-corba-integration</artifactId>
   <name>TomEE :: Container :: Corba Integration</name>
 
-  <!-- FIXME: Separate Project to build it with Java 8 separately -->
-  <artifactId>openejb-corba-integration</artifactId>
+  <dependencies>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>openejb-core</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>openejb-loader</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+  </dependencies>
 
-  <modules>
-    <module>openejb-corba-integration-api</module>
-    <module>openejb-corba-integration-impl</module>
-  </modules>
 </project>
