+++ container/openejb-corba-integration/src/main/java/org/apache/openejb/corba/client/CorbasProviderInstaller.java
@@ -14,9 +14,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.openejb.corba;
+package org.apache.openejb.corba.client;
 
-public interface ORBAdapter {
+import org.apache.openejb.loader.SystemInstance;
+import org.apache.openejb.observer.Observes;
+import org.apache.openejb.observer.event.ObserverAdded;
+import org.apache.openejb.spi.client.corba.CorbasProvider;
 
-    Object getORB();
+public class CorbasProviderInstaller {
+
+    public void install(@Observes final ObserverAdded added) {
+        if (added.getObserver() == this) {
+            SystemInstance.get().setComponent(CorbasProvider.class, new CorbasProviderImpl());
+        }
+    }
 }
