+++ container/openejb-corba-integration/src/main/resources/META-INF/service-jar.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  Licensed to the Apache Software Foundation (ASF) under one or more
+  contributor license agreements.  See the NOTICE file distributed with
+  this work for additional information regarding copyright ownership.
+  The ASF licenses this file to You under the Apache License, Version 2.0
+  (the "License"); you may not use this file except in compliance with
+  the License.  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+-->
+
+<!-- $Rev$ $Date$ -->
+
+<ServiceJar>
+
+    <!--
+    #===================================================================#
+    #       D E F A U L T    O P E N E J B    S E R V I C E S           #
+    #===================================================================#
+    #
+    #
+    #  This file defines all the default configurations of containers,
+    #  factories, managers and services that come with OpenEJB.
+    #
+    #  Can contain numerous configurations of the standard services
+    #  making it easy for users to simply name the one they want in
+    #  their configuration files.
+    #
+    #  Over time this will be become a library of default
+    #  configurations
+    -->
+
+    <!--
+    # ================================================
+    # Default ORB
+    # ================================================
+    -->
+    <ServiceProvider
+            id="Default ORB"
+            service="Resource"
+            types="org.apache.openejb.spi.corba.ORBAdapter, ORBAdapter"
+            factory-name="create"
+            class-name="org.apache.openejb.corba.ORBFactoryProviderImpl">
+
+    </ServiceProvider>
+
+</ServiceJar>
